main_pastel_theme.py
############################################# Enhanced Face Recognition System with Pastel Theme ################################################

import tkinter as tk
from tkinter import ttk
from tkinter import messagebox as mess
from tkinter import PhotoImage
from PIL import Image, ImageTk, ImageDraw, ImageFilter
import tkinter.simpledialog as tsd
import cv2, os
import csv
import numpy as np
from PIL import Image
import pandas as pd
import datetime
import time
import threading
import warnings
warnings.filterwarnings("ignore")

# Import semi-transparent frame classes
from semi_transparent_frames import SemiTransparentFrame, GlassFrame, AnimatedGlassFrame, FrameFactory, create_layered_background

# GPU acceleration imports
import torch
import torch.nn.functional as F
from torchvision import transforms
import torchvision.transforms as T

# Email imports
import smtplib
from email.mime.multipart import MIMEMultipart
from email.mime.text import MIMEText
from email.mime.base import MIMEBase
from email import encoders

# Global variables
global key
key = ""

# Main window with enhanced visuals and pastel theme
window = tk.Tk()
window.geometry("1600x750")
window.resizable(True, False)
window.title("🚀 Enhanced Face Recognition Attendance System")
window.configure(background="#f8f9fa")

# Function declarations (GPUManager, EnhancedFaceDetector, etc.)
# Place previously implemented functions here

# Example GUI element setup
title_frame = GlassFrame(window, x=0, y=0, width=1600, height=150, alpha=0.95, base_color=(255, 250, 240), accent_color=(255, 218, 185))
message = tk.Label(title_frame.canvas, text="Welcome to the Enhanced Face Recognition System", fg="#4a5568", bg="#fff8dc", font=("Helvetica", 24, " bold "))
message.place(relx=0.5, rely=0.5, anchor="center")

# Run main loop
window.mainloop()

#
# Implement the rest of the main loops and event bindings as needed.
#
